Rename dom-for in file paths to domFor
